fix: MyInvitation 패키지 구조 개선
- ListDetail 패키지 구조 분리
