+++ android/feature/myinvitation/src/main/kotlin/com/andlife/myinvitation/MyInvitationGraph.kt
@@ -14,8 +14,8 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.andlife.domain.util.RefreshEventHub
-import com.andlife.myinvitation.screen.MyInvitationListDetailRoute
 import com.andlife.myinvitation.screen.detail.MyInvitationDetailRoute
+import com.andlife.myinvitation.screen.listdetail.MyInvitationListDetailRoute
 import com.andlife.myinvitation.viewmodel.MyInvitationDetailViewModel
 import com.andlife.myinvitation.viewmodel.MyInvitationGuestBookViewModel
 import com.andlife.myinvitation.viewmodel.MyInvitationViewModel
