+++ android/feature/myinvitation/src/main/kotlin/com/andlife/myinvitation/screen/list/MyInvitationScreen.kt
@@ -1,4 +1,4 @@
-package com.andlife.myinvitation.screen
+package com.andlife.myinvitation.screen.list
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.background
@@ -62,16 +62,9 @@ import com.andlife.designsystem.theme.NachoTheme
 import com.andlife.domain.model.auth.AuthState
 import com.andlife.domain.model.invitation.SortDirection
 import com.andlife.model.invitation.InvitationSummaryUiModel
-import com.andlife.myinvitation.MyInvitationDetail
-import com.andlife.myinvitation.MyInvitationPlaceholder
 import com.andlife.myinvitation.model.MyInvitationSideEffect
 import com.andlife.myinvitation.model.MyInvitationUiEvent
 import com.andlife.myinvitation.model.MyInvitationUiState
-import com.andlife.myinvitation.screen.detail.MyInvitationDetailRoute
-import com.andlife.myinvitation.screen.detail.MyInvitationPlaceholderScreen
-import com.andlife.myinvitation.viewmodel.MyInvitation2PaneViewModel
-import com.andlife.myinvitation.viewmodel.MyInvitationDetailViewModel
-import com.andlife.myinvitation.viewmodel.MyInvitationGuestBookViewModel
 import com.andlife.myinvitation.viewmodel.MyInvitationViewModel
 import com.andlife.ui.R
 import com.andlife.ui.component.GenericTabRow
@@ -81,153 +74,11 @@ import com.andlife.ui.component.listitem.InvitationListItem
 import com.andlife.ui.component.listitem.MenuItem
 import com.andlife.ui.component.loading.InvitationLoadingIndicator
 import com.andlife.ui.component.paging.PagingStateContent
-import com.andlife.ui.scope.DetailPaneScopedViewModel
-import com.andlife.ui.util.LocalNavigationSuiteState
 import com.andlife.ui.util.collectWithLifecycle
-import com.andlife.ui.util.isNavigationBar
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.launch
 import com.andlife.designsystem.R as designR
-
-@Composable
-fun MyInvitationListDetailRoute(
-    snackbarHostState: SnackbarHostState,
-    onNavigateToLogin: () -> Unit,
-    onNavigateToCreate: () -> Unit,
-    onNavigateToEditInvitation: (Long) -> Unit,
-    onNavigateToEditCard: (Long) -> Unit,
-    onNavigateToCreateCard: (Long) -> Unit,
-    onNavigateToCreateThanksCard: (Long) -> Unit,
-    onNavigateToUpdateThanksCard: (Long) -> Unit,
-    viewModel: MyInvitation2PaneViewModel = hiltViewModel()
-) {
-    val uiState by viewModel.initialInvitationId.collectAsStateWithLifecycle()
-
-    MyInvitationListDetailScreen(
-        initialInvitationId = uiState,
-        snackbarHostState = snackbarHostState,
-        onNavigateToLogin = onNavigateToLogin,
-        onNavigateToCreate = onNavigateToCreate,
-        onNavigateToEditInvitation = onNavigateToEditInvitation,
-        onNavigateToEditCard = onNavigateToEditCard,
-        onNavigateToCreateCard = onNavigateToCreateCard,
-        onNavigateToCreateThanksCard = onNavigateToCreateThanksCard,
-        onNavigateToUpdateThanksCard = onNavigateToUpdateThanksCard,
-        onNavigateToDetail = viewModel::setInitialInvitationId,
-    )
-}
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-@Composable
-private fun MyInvitationListDetailScreen(
-    initialInvitationId: Long?,
-    snackbarHostState: SnackbarHostState,
-    onNavigateToDetail: (Long) -> Unit,
-    onNavigateToLogin: () -> Unit,
-    onNavigateToCreate: () -> Unit,
-    onNavigateToEditInvitation: (Long) -> Unit,
-    onNavigateToEditCard: (Long) -> Unit,
-    onNavigateToCreateCard: (Long) -> Unit,
-    onNavigateToCreateThanksCard: (Long) -> Unit,
-    onNavigateToUpdateThanksCard: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-
-    val suiteState = LocalNavigationSuiteState.current
-
-    val navSuiteType =
-        NavigationSuiteScaffoldDefaults.navigationSuiteType(currentWindowAdaptiveInfo())
-
-    val listDetailNavigator = rememberListDetailPaneScaffoldNavigator()
-    val coroutineScope = rememberCoroutineScope()
-
-    var myInvitationRoute by remember {
-        val route = initialInvitationId?.let { MyInvitationDetail(id = it) } ?: MyInvitationPlaceholder
-        mutableStateOf(route)
-    }
-
-    fun onInvitationClickShowDetailPane(id: Long) {
-        onNavigateToDetail(id)
-        myInvitationRoute = MyInvitationDetail(id)
-        coroutineScope.launch {
-            if (navSuiteType.isNavigationBar && suiteState.currentValue == NavigationSuiteScaffoldValue.Visible) suiteState.hide()
-            listDetailNavigator.navigateTo(ListDetailPaneScaffoldRole.Detail)
-        }
-    }
-
-    ThreePaneScaffoldPredictiveBackHandler(
-        listDetailNavigator,
-        BackNavigationBehavior.PopUntilScaffoldValueChange,
-    )
-
-    NavigableListDetailPaneScaffold(
-        navigator = listDetailNavigator,
-        listPane = {
-            AnimatedPane {
-                MyInvitationRoute(
-                    snackbarHostState = snackbarHostState,
-                    onNavigateToDetail = { onInvitationClickShowDetailPane(it) },
-                    onNavigateToLogin = onNavigateToLogin,
-                    onNavigateToCreate = onNavigateToCreate,
-                    selectedInvitationId = initialInvitationId,
-                    shouldHighlightSelected = listDetailNavigator.isDetailPaneVisible(),
-                )
-            }
-        },
-        detailPane = {
-            AnimatedPane {
-                AnimatedContent(myInvitationRoute) { route ->
-                    when (route) {
-                        is MyInvitationDetail -> {
-                            DetailPaneScopedViewModel {
-                                MyInvitationDetailRoute(
-                                    selectedId = route.id,
-                                    onNavigateBack = {
-                                        coroutineScope.launch {
-                                            if (navSuiteType.isNavigationBar && suiteState.currentValue == NavigationSuiteScaffoldValue.Hidden) suiteState.show()
-                                            listDetailNavigator.navigateBack()
-                                        }
-                                    },
-                                    onNavigateToLogin = onNavigateToLogin,
-                                    onNavigateToEditCard = onNavigateToEditCard,
-                                    onNavigateToEditInvitation = onNavigateToEditInvitation,
-                                    onNavigateToCreateCard = onNavigateToCreateCard,
-                                    onNavigateToCreateThanksCard = onNavigateToCreateThanksCard,
-                                    onNavigateToUpdateThanksCard = onNavigateToUpdateThanksCard,
-                                    showBackButton = !listDetailNavigator.isListPaneVisible(),
-                                    viewModel = hiltViewModel<MyInvitationDetailViewModel, MyInvitationDetailViewModel.Factory>(
-                                        key = "myDetail ${route.id}"
-                                    ) { factory ->
-                                        factory.create(route.id)
-                                    },
-                                    guestBookViewModel = hiltViewModel<MyInvitationGuestBookViewModel, MyInvitationGuestBookViewModel.Factory>(
-                                        key = "myDetail guestBook ${route.id}"
-                                    ) { factory ->
-                                        factory.create(route.id)
-                                    }
-                                )
-                            }
-                        }
-
-                        MyInvitationPlaceholder -> {
-                            MyInvitationPlaceholderScreen()
-                        }
-                    }
-                }
-            }
-        },
-        modifier = modifier.background(NachoTheme.colorScheme.backgroundPrimary),
-    )
-}
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isListPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.List] == PaneAdaptedValue.Expanded
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isDetailPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.Detail] == PaneAdaptedValue.Expanded
 
 @Composable
 fun MyInvitationRoute(
